Allow users to disable regular output completely
Especially useful when printing logs to stdout with log-target=stdout.
